Parameters

Parameters

Arguments

aspectRatio
integer(1). Aspect ratio.

assay
vector(1). Name or index of count matrix slotted in assays(). When passing in a string, the name must be defined in assayNames().

borderColor
character(1) or NULL. Border color. Disabled by default for improved aesthetics.

color
ScaleDiscrete. Desired ggplot2 color scale. Must supply discrete values. When set NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_color_manual().

To set the discrete color palette globally, use:

options(acid.color.discrete = ggplot2::scale_color_viridis_d())

countsAxisLabel
character(1). Counts axis label.

dark
logical(1). Plot against a dark background using the theme_midnight() ggplot2 theme.

dimsUse
integer. Vector of length 2 that denotes the columns from the reduced dimension matrix to use for centerX and centerY column calculations. Defaults the first and second dimensions.

direction
character(1). Plot "both", "up", or "down" directions.

expression
character(1). Calculation to apply. Uses match.arg() internally and defaults to the first argument in the character vector.

fill
ggproto/ScaleDiscrete. Desired ggplot2 fill scale. Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_fill_manual().

To set the discrete fill palette globally, use:

options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())

flip
logical(1). Flip x and y axes. Recommended for plots containing many samples.

gene2symbol
Gene2Symbol. Gene-to-symbol mappings. Must contain geneID and geneName columns. See Gene2Symbol for more information.

genes
character. Gene identifiers. It is considered better practice to input the stable gene identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene symbols (e.g. "TSPN6"), if possible.

geom
character(1). Plot type. Uses match.arg() internally and defaults to the first argument in the character vector.

headerLevel
integer(1) (1-7). Markdown header level.

interestingGroups
character. Groups of interest that define the samples. If left unset, defaults to sampleName.

label
logical(1). Superimpose sample text labels on the plot.

labelSize
integer(1). Size of the text label.

legend
logical(1). Show plot legend.

level
character(1). Return as genes or transcripts.

limit
numeric(1). Threshold to denote on the plot, using a dashed line.

max
numeric(1). Recommended maximum value cutoff.

min
numeric(1). Recommended minimum value cutoff.

minCounts
integer(1). Minimum number of counts per gene in the count matrix.

n
integer(1). Number to include.

ntop
integer(1). Number of top genes to label.

object
Object.

perMillion
logical(1). Display as counts per million.

perSample
logical(1). Visualize the distributions per sample.

plotlist
list. List containing ggplot objects.

pointAlpha
numeric(1) (0-1). Alpha transparency level. Useful when there many points in the dataset (e.g. single-cell data), and some points can be masked.

pointColor
character(1). Default point color for the plot.

pointScalar
integer(1). Default point size for the plot.

pointsAsNumbers
logical(1). Plot the points as numbers (TRUE) or dots (FALSE).

pointSize
numeric(1). Point size for dots in the plot.

return
character(1). Return type. Uses match.arg() internally and defaults to the first argument in the character vector.

scale
character(1). Whether the values should be centered and scaled in either the row or column direction ("row", "column"), or remain unscaled ("none").

subtitle
character(1). Plot subtitle.

title
character(1). Plot title.

trans
character(1). Name of the axis scale transformation to apply.

For more information:

help(topic = "scale_x_continuous", package = "ggplot2")

trendline
logical(1). Include trendline on plot.

...
Additional arguments.

Value

No value.